функции установки статуса формы и фильтра вынес в togglePageStstus.js
