+++ js/form.js
@@ -1,12 +1,6 @@
-const adForm = document.querySelector('.ad-form');
+import {setAdFormOff, setAdFormOn} from './togglePageStstus.js';
 
-const setAdFormDisabled = (flagDisabled) => {
-  adForm.classList.toggle('.ad-form--disabled');
-  [...adForm.children].forEach((item) => {
-    item.disabled = flagDisabled;
-  });
-};
-const setAdFormOff = () => setAdFormDisabled(true);
-const setAdFormOn = () => setAdFormDisabled(false);
+setAdFormOn();
+setAdFormOff();
 
 export {setAdFormOff, setAdFormOn};
