+++ js/main.js
@@ -1,13 +1,5 @@
 import {creatListAdvertisement} from './data.js';
 import './card.js';
-import {setAdFormOff, setAdFormOn} from './form.js';
-import {setMapFilterOff, setMapFilterOn} from './filter.js';
+import './form.js';
+import './filter.js';
 creatListAdvertisement();
-
-setAdFormOn();
-setAdFormOff();
-
-setMapFilterOn();
-setMapFilterOff();
-
-
